Remove linters error.
